remove extra script, add DFT code to original
